+++ src/main/java/com/ll/netmong/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.ll.netmong.member.repository;
+package com.ll.netmong.domain.member.repository;
 
-import com.ll.netmong.member.entity.Member;
+import com.ll.netmong.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
