+++ src/main/java/com/ll/netmong/domain/member/service/MemberService.java
@@ -1,7 +1,7 @@
-package com.ll.netmong.member.service;
+package com.ll.netmong.domain.member.service;
 
-import com.ll.netmong.member.entity.Member;
-import com.ll.netmong.member.repository.MemberRepository;
+import com.ll.netmong.domain.member.entity.Member;
+import com.ll.netmong.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
